feat: Add variables
