+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Awesome</title>
-    <script src="index.js" defer></script>
+    <script src="index.js" type="module" defer></script>
     <link rel="stylesheet" href="index.css">
 </head>
 <body>
