+++ modules/variables.js
@@ -1,0 +1,9 @@
+const dateTime = document.querySelector('.date-time');
+const list = document.getElementById('list');
+const addList = document.getElementById('add');
+const contact = document.getElementById('contact');
+const displaySection = document.getElementById('book-list');
+const inputSection = document.querySelector('.book-input');
+const contactSection = document.querySelector('.contact-section');
+
+export { dateTime, list, addList, contact, displaySection, inputSection, contactSection };
